try to parse all requests as json
Try to parse all requests as json and fallback to returning text if
unparsable.

In case the body is unparsable but content-type is json we throw an
error.

Also add tests for various cases.
