+++ src/api.mjs
@@ -1,6 +1,5 @@
 // @format
 import Ajv from "ajv";
-import fetch from "cross-fetch";
 import {
   exit as exitMsg,
   https,
@@ -13,6 +12,7 @@ import logger from "./logger.mjs";
 import { ValidationError, NotImplementedError } from "./errors.mjs";
 import { translate } from "./eth.mjs";
 import { endpointStore } from "./endpoint_store.mjs";
+import { request } from "./request.mjs";
 
 const log = logger("api");
 const ajv = new Ajv();
@@ -39,48 +39,6 @@ function validate(value) {
   }
 
   return true;
-}
-
-export async function request(url, method, body, headers, signal) {
-  let options = {
-    method,
-  };
-
-  if (body) {
-    options.body = body;
-  }
-  if (headers) {
-    options.headers = headers;
-  }
-  if (signal) {
-    options.signal = signal;
-  }
-
-  // NOTE: We let `fetch` throw. Error must be caught on `request` user level.
-  const results = await fetch(url, options);
-  const answer = await results.text();
-
-  if (results.status >= 400) {
-    throw new Error(
-      `Request to url "${url}" with method "${method}" and body "${JSON.stringify(
-        body
-      )}" unsuccessful with status: ${results.status} and answer: "${answer}"`
-    );
-  }
-
-  if (results.headers.get("Content-Type")?.toLowerCase().includes("json")) {
-    let data;
-    try {
-      data = JSON.parse(answer);
-    } catch (err) {
-      throw new Error(
-        `Encountered error when trying to parse JSON body result: "${answer}", error: "${err.toString()}"`
-      );
-    }
-    return data;
-  }
-
-  return answer;
 }
 
 // NOTE: `AbortSignal.timeout` isn't yet supported:
